make_lines

Makes the lines in the corner of the base.

make_lines(
  heightmap,
  basedepth = 0,
  linecolor = "grey20",
  zscale = 1,
  alpha = 1,
  linewidth = 2,
  solid = TRUE
)

Arguments

heightmap
A two-dimensional matrix, where each entry in the matrix is the elevation at that point. All points are assumed to be evenly spaced.

basedepth
Default `0`.

linecolor
Default `grey40`.

zscale
Default `1`. The ratio between the x and y spacing (which are assumed to be equal) and the z axis. For example, if the elevation levels are in units of 1 meter and the grid values are separated by 10 meters, `zscale` would be 10.

alpha
Default `1`. Transparency.

linewidth
Default `2`. Linewidth

solid
Default `TRUE`. Whether it's solid or water.