make_waterlines

Makes the edge lines of

make_waterlines(heightmap, waterdepth = 0, linecolor = "grey40",
  zscale = 1, alpha = 1, linewidth = 2, antialias = FALSE)

Arguments

heightmap
A two-dimensional matrix, where each entry in the matrix is the elevation at that point. All points are assumed to be evenly spaced.

waterdepth
Default `0`.

linecolor
Default `grey40`.

zscale
Default `1`. The ratio between the x and y spacing (which are assumed to be equal) and the z axis. For example, if the elevation levels are in units of 1 meter and the grid values are separated by 10 meters, `zscale` would be 10.

alpha
Default `1`. Transparency of lines.

linewidth
Default `2`. Water line width.

antialias
Default `FALSE`.