plot_gg()

Plots a ggplot object in 3D

plot_gg(ggplotobj, heighttype = "fill", width = 3, height = 3,
  units = c("in", "cm", "mm"), scale = 150, pointcontract = 1,
  raytrace = TRUE, sunangle = 315, anglebreaks = seq(30, 40, 0.1),
  ...)

Arguments

ggplotobj
Hillshade/image to be added to 3D surface map.

heighttype
Default `fill`. Whether the `fill` or `color` aesthetic should be used for height values.

width
Default `3`. Width of ggplot, in `units`

height
Default `3`. Height of ggplot, in `units`

units
Default `in`. One of c("in", "cm", "mm").

scale
Default `150`. Amount of vertical scaling.

raytrace
Default `FALSE`. Whether to add a raytraced layer.

sunangle
Default `315` (NW). If raytracing, the angle (in degrees) around the matrix from which the light originates.

linewidth
Default `2`. Width of the edge lines in the scene.

maxsearch
Default `100`. If raytracing, the maximum distance that the system should propogate rays to check. For longer

lambert
Default `TRUE`. If raytracing, changes the intensity of the light at each point based proportional to the dot product of the ray direction and the surface normal at that point. Zeros out all values directed away from the ray.

multicore
Default `FALSE`. If raytracing and `TRUE`, multiple cores will be used to compute the shadow matrix. By default, this uses all cores available, unless the user has set `options("cores")` in which the multicore option will only use that many cores.

progbar
Default `TRUE`. If raytracing and `FALSE`, turns off progress bar.

Value

4-layer RGBA array.

Examples

#dum dum